% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_model.R, R/generics.R
\name{define_model.data.frame}
\alias{define_model.data.frame}
\alias{define_model.infer_model}
\alias{rel}
\alias{compare_by}
\alias{cont_tab}
\alias{prop_model}
\alias{pairwise}
\alias{selected_vars}
\alias{one_sample}
\alias{define-model}
\alias{define_model}
\title{Define a model for statistical inference}
\usage{
\method{define_model}{data.frame}(.x, to_analyze, ...)

\method{define_model}{infer_model}(.x, data = NULL, ...)

rel(x, resp, ...)

compare_by(x, grp, ...)

cont_tab(ind, dep, ...)

prop_model(x, ...)

pairwise(..., .dir = "leq")

selected_vars(...)

one_sample(var, ...)

define_model(.x, ...)
}
\arguments{
\item{.x}{A data frame, or missing for lexical scoping.}

\item{to_analyze}{A model declaration from \code{rel()}, \code{cont_tab()},
\code{prop_model()}, or \code{pairwise()}.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A \code{stat_infer_pipeline} object.
}
\description{
\code{define_model()} initializes the pipeline for statistical inference.
}
