% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{display}
\alias{display}
\title{Display the test output}
\usage{
display(.x, ...)
}
\arguments{
\item{.x}{A \code{stat_infer_out} object.}

\item{...}{Additional arguments passed to methods.}
}
\description{
\code{display()} prints a formatted summary to the console. Returns invisibly.
}
