% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-extra_args.R
\name{extra_args}
\alias{extra_args}
\title{Supplied arguments declaration}
\usage{
extra_args(...)
}
\description{
Use one-sided formula \code{~n} for no default, two-sided \code{n ~ 1000L} for default.
}
\keyword{internal}
