% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_model.R
\name{get_model_id}
\alias{get_model_id}
\title{Extract the model_id from a pipeline object}
\usage{
get_model_id(.x)
}
\arguments{
\item{.x}{A \code{model_to_analyze} or \code{test_lazy} object.}
}
\description{
Extract the model_id from a pipeline object
}
