% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferlystat-package.R
\docType{package}
\name{inferlystat}
\alias{inferlystat-package}
\alias{inferlystat}
\title{inferlystat: Composable and Tidy Grammar for Statistical Inference}
\description{
Provides a composable, pipe-based grammar for statistical inference in R. The package defines a unified domain-specific language for expressing statistical workflows, including data specification, estimands, hypotheses, test statistics, and evaluation, using tidyverse conventions. It supports classical, resampling-based, and Bayesian inference paradigms, and returns tidy outputs suitable for downstream analysis and reporting.
}
\author{
\strong{Maintainer}: Joshua Marie \email{joshua.marie.k@gmail.com}

}
\keyword{internal}
