% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-bayes.R
\name{method_bayes}
\alias{method_bayes}
\title{Bayesian method declaration}
\usage{
method_bayes(method_name, args = NULL)
}
\arguments{
\item{method_name}{A string naming the method, e.g. \code{"permute"}, \code{"boot"}.}

\item{args}{An \code{extra_args} object declaring method-specific arguments.}
}
\description{
Bayesian method declaration
}
