% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot-test}
\alias{plot-test}
\alias{plot_test}
\title{Plot the test output}
\usage{
plot_test(.x, ...)
}
\arguments{
\item{.x}{A \code{stat_infer_out} object.}

\item{...}{Additional arguments passed to methods.}
}
\description{
\code{plot_test()} produces a plot for a single-test pipeline. Errors
informatively if more than one test is registered.
}
