% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-replicate.R
\name{method_replicate}
\alias{method_replicate}
\title{Declare a replication method}
\usage{
method_replicate(method_name, args = NULL)
}
\arguments{
\item{method_name}{A string naming the method, e.g. \code{"permute"}, \code{"boot"}.}

\item{args}{An \code{extra_args} object declaring method-specific arguments.}
}
\description{
Declare a replication method
}
